requirements.txt
aiohttp==3.9.5
Faker==26.0.0
python-dotenv==1.0.1